+++ src/ProtectedEvaluator.ts
@@ -1,9 +1,8 @@
-import {Utils as FormioUtils} from '@formio/js';
+import {Utils as FormioUtils, DefaultEvaluator} from '@formio/js';
 import Interpreter from '@formio/js-interpreter';
 
-const baseEvaluator = (FormioUtils.Evaluator as any).evaluator;
-const baseEvaluate = (FormioUtils.Evaluator as any).evaluate;
-const DefaultEvaluator = (FormioUtils as any).DefaultEvaluator;
+const baseEvaluator = FormioUtils.Evaluator.evaluator;
+const baseEvaluate = FormioUtils.Evaluator.evaluate;
 
 export interface IEvaluator {
   noeval?: boolean;